+++ CHANGELOG.md
@@ -31,10 +31,14 @@
 
 ### Doc
 - remove stability warning
+- image size
 
 ### Feat
 - shorten time interval for liveness and readiness probe ([#12](https://github.com/torbendury/kube-networkpolicy-denier/issues/12))
 - **docker:** minify image size ([#15](https://github.com/torbendury/kube-networkpolicy-denier/issues/15))
+
+### Refactor
+- **core:** extract admission review functionality for better testing capability
 
 
 <a name="kube-networkpolicy-denier-0.1.0"></a>